requirements/development.txt
-r base.txt

# Development tools
django-debug-toolbar==4.2.0
ipython==8.20.0
black==23.12.1
flake8==7.0.0
pytest==7.4.4
pytest-django==4.7.0